chore: change worksheet protection file location
