+++ packages/sheets-ui/src/commands/commands/worksheet-protection.command.ts
@@ -15,124 +15,11 @@
  */
 
 import type { ICommand, Workbook } from '@univerjs/core';
-import type { IWorksheetProtectionRule } from '@univerjs/sheets';
 
 import { CommandType, ICommandService, IUndoRedoService, IUniverInstanceService } from '@univerjs/core';
 
 import { UniverType } from '@univerjs/protocol';
-import { AddWorksheetProtectionMutation, DeleteWorksheetProtectionMutation, SetWorksheetProtectionMutation, WorksheetProtectionRuleModel } from '@univerjs/sheets';
-
-export interface IAddWorksheetProtectionParams {
-    unitId: string;
-    rule: IWorksheetProtectionRule;
-}
-
-export interface IDeleteWorksheetProtectionParams {
-    unitId: string;
-    subUnitId: string;
-    rule: IWorksheetProtectionRule;
-}
-
-export interface ISetWorksheetProtectionParams {
-    permissionId: string;
-    rule: IWorksheetProtectionRule;
-    oldRule: IWorksheetProtectionRule;
-}
-
-export const AddWorksheetProtectionCommand: ICommand<IAddWorksheetProtectionParams> = {
-    type: CommandType.COMMAND,
-    id: 'sheet.command.add-worksheet-protection',
-    async handler(accessor, params) {
-        if (!params) {
-            return false;
-        }
-        const commandService = accessor.get(ICommandService);
-        const undoRedoService = accessor.get(IUndoRedoService);
-        const { rule, unitId } = params;
-        const subUnitId = rule.subUnitId;
-
-        const result = await commandService.executeCommand(AddWorksheetProtectionMutation.id, {
-            unitId,
-            rule,
-            subUnitId: rule.subUnitId,
-        });
-
-        if (result) {
-            const redoMutations = [{ id: AddWorksheetProtectionMutation.id, params: { unitId, rule, subUnitId: rule.subUnitId } }];
-            const undoMutations = [{ id: DeleteWorksheetProtectionMutation.id, params: { unitId, subUnitId } }];
-            undoRedoService.pushUndoRedo({
-                unitID: unitId,
-                redoMutations,
-                undoMutations,
-            });
-        }
-
-        return true;
-    },
-};
-
-export const DeleteWorksheetProtectionCommand: ICommand<IDeleteWorksheetProtectionParams> = {
-    type: CommandType.COMMAND,
-    id: 'sheet.command.delete-worksheet-protection',
-    handler(accessor, params) {
-        if (!params) {
-            return false;
-        }
-        const commandService = accessor.get(ICommandService);
-        const undoRedoService = accessor.get(IUndoRedoService);
-        const { rule, unitId, subUnitId } = params;
-
-        commandService.executeCommand(DeleteWorksheetProtectionMutation.id, {
-            unitId,
-            subUnitId,
-        });
-
-        const redoMutations = [{ id: DeleteWorksheetProtectionMutation.id, params: { unitId, subUnitId } }];
-        const undoMutations = [{ id: AddWorksheetProtectionMutation.id, params: { unitId, rule, subUnitId } }];
-        undoRedoService.pushUndoRedo({
-            unitID: unitId,
-            redoMutations,
-            undoMutations,
-        });
-
-        return true;
-    },
-};
-
-export const SetWorksheetProtectionCommand: ICommand<ISetWorksheetProtectionParams> = {
-    type: CommandType.COMMAND,
-    id: 'sheet.command.set-worksheet-protection',
-    async handler(accessor, params) {
-        if (!params) {
-            return false;
-        }
-        const commandService = accessor.get(ICommandService);
-        const undoRedoService = accessor.get(IUndoRedoService);
-
-        const { rule, permissionId, oldRule } = params;
-        const { unitId, subUnitId } = rule;
-
-        const newRule = { ...rule, permissionId };
-
-        const result = await commandService.executeCommand(SetWorksheetProtectionMutation.id, {
-            unitId,
-            subUnitId,
-            newRule,
-        });
-
-        if (result) {
-            const redoMutations = [{ id: SetWorksheetProtectionMutation.id, params: { unitId, subUnitId, newRule } }];
-            const undoMutations = [{ id: SetWorksheetProtectionMutation.id, params: { unitId, subUnitId, rule: oldRule } }];
-            undoRedoService.pushUndoRedo({
-                unitID: unitId,
-                redoMutations,
-                undoMutations,
-            });
-        }
-
-        return true;
-    },
-};
+import { AddWorksheetProtectionCommand, DeleteWorksheetProtectionCommand, WorksheetProtectionRuleModel } from '@univerjs/sheets';
 
 export const DeleteWorksheetProtectionFormSheetBarCommand: ICommand = {
     type: CommandType.COMMAND,
